Refactor existing shell scripts & moving scripts around

+++ core/tabs/applications-setup/office-suite-setup.sh
@@ -3,8 +3,8 @@
 . ../common-script.sh
 
 install_onlyoffice() {
-    printf "%b\n" "${YELLOW}Installing Only Office..${RC}."
     if ! command_exists onlyoffice-desktopeditors; then
+        printf "%b\n" "${YELLOW}Installing Only Office..${RC}."
         case "$PACKAGER" in
             apt-get|nala)
                 curl -O https://download.onlyoffice.com/install/desktop/editors/linux/onlyoffice-desktopeditors_amd64.deb
@@ -27,8 +27,8 @@ install_onlyoffice() {
 }
 
 install_libreoffice() {
-    printf "%b\n" "${YELLOW}Installing Libre Office...${RC}"
     if ! command_exists libreoffice; then
+        printf "%b\n" "${YELLOW}Installing Libre Office...${RC}"
         case "$PACKAGER" in
         apt-get|nala)
             "$ESCALATION_TOOL" "$PACKAGER" install -y libreoffice-core
@@ -51,8 +51,8 @@ install_libreoffice() {
 }
 
 install_wpsoffice() {
-    printf "%b\n" "${YELLOW}Installing WPS Office...${RC}"
     if ! command_exists com.wps.Office; then
+        printf "%b\n" "${YELLOW}Installing WPS Office...${RC}"
         case "$PACKAGER" in
             pacman)
                 "$AUR_HELPER" -S --noconfirm wps-office
@@ -69,8 +69,9 @@ install_wpsoffice() {
 
 # needs to be updated every year for latest version
 install_freeoffice() {  
-	printf "%b\n" "${YELLOW}Installing Free Office...${RC}"
-    case "$PACKAGER" in
+    if ! command_exists softmaker-freeoffice-2024 freeoffice softmaker; then
+        printf "%b\n" "${YELLOW}Installing Free Office...${RC}"
+        case "$PACKAGER" in
         apt-get|nala)
             curl -O https://www.softmaker.net/down/softmaker-freeoffice-2024_1218-01_amd64.deb
             "$ESCALATION_TOOL" dpkg -i softmaker-freeoffice-2024_1218-01_amd64.deb
@@ -91,12 +92,15 @@ install_freeoffice() {
         *)
             printf "%b\n" "${RED}The script does not support your Distro. Install manually..${RC}"
             ;;
-    esac
+        esac
+    else
+        printf "%b\n" "${GREEN}Free Office is already installed.${RC}"
+    fi
 }
 
 install_evince() {
-    printf "%b\n" "${YELLOW}Installing Evince...${RC}"
     if ! command_exists evince; then
+        printf "%b\n" "${YELLOW}Installing Evince...${RC}"
         case "$PACKAGER" in
             pacman)
                 "$ESCALATION_TOOL" "$PACKAGER" -S --noconfirm evince
@@ -111,8 +115,8 @@ install_evince() {
 }
 
 install_okular() {
-    printf "%b\n" "${YELLOW}Installing Evince...${RC}"
     if ! command_exists okular; then
+        printf "%b\n" "${YELLOW}Installing Okular...${RC}"
         case "$PACKAGER" in
             pacman)
                 "$ESCALATION_TOOL" "$PACKAGER" -S --noconfirm okular
@@ -127,8 +131,8 @@ install_okular() {
 }
 
 install_pdfstudioviewer() {
-    printf "%b\n" "${YELLOW}Installing PDF Studio Viewer...${RC}"
     if ! command_exists pdfstudioviewer2024/pdfstudioviewer2024; then
+        printf "%b\n" "${YELLOW}Installing PDF Studio Viewer...${RC}"
         curl -O https://download.qoppa.com/pdfstudioviewer/PDFStudioViewer_linux64.sh
         "$ESCALATION_TOOL" chmod +x PDFStudioViewer_linux64.sh
         if sh PDFStudioViewer_linux64.sh; then
@@ -143,8 +147,8 @@ install_pdfstudioviewer() {
 }
 
 install_pdfstudio() {
-    printf "%b\n" "${YELLOW}Installing PDF Studio Viewer...${RC}"
     if ! command_exists pdfstudio2024/pdfstudio2024; then
+        printf "%b\n" "${YELLOW}Installing PDF Studio...${RC}"
         curl -O https://download.qoppa.com/pdfstudio/PDFStudio_linux64.sh
         "$ESCALATION_TOOL" chmod +x PDFStudio_linux64.sh
         if sh PDFStudio_linux64.sh; then
@@ -157,7 +161,6 @@ install_pdfstudio() {
         printf "%b\n" "${GREEN}PDF Studio is already installed.${RC}"
     fi
 }
-
 
 officeSuiteSetup() {
     clear
@@ -175,7 +178,7 @@ officeSuiteSetup() {
     printf "%b\n" "7. PDF Studio Viewer"
     printf "%b\n" "8. PDF Studio (Paid Software)"
     printf "%b\n" "----------------------------"
-    printf "%b\n"  "Enter your choices (e.g., 1 3 5): "
+    printf "%b"  "Enter your choices (e.g., 1 3 5): "
     read -r choice
     for ch in $choice; do
         case $ch in
